move player movement restrict logic to common/maplib
