+++ common/maplib/package.json
@@ -1,4 +1,7 @@
 {
   "name": "hitcon-online-maplib",
-  "version": "0.0.1"
+  "version": "0.0.1",
+  "dependencies": {
+    "config": "^3.3.6"
+  }
 }
